Initialized QuaLiKiz-pythontools package skeleton
By moving to this new structure, the Python tools can be installed
just like a regular Python package, e.g. using pip. This also sets
up the framework to create a Command Line Interface, or CLI. Note
that this only is the skeleton! After this move, all modules will
be broken, and will have to be fixed.