Module to convert QuaLiKiz output to netCDF4/xarray
netCDF4 is a more efficient storage format than raw ASCII
These scripts can load the ASCII files in debug, output, and
output/primitive and convert them to the xarray DataSets,
which can be easily stored to disk as netCDF4 files.
Additionally, the DataSet can be recast to a hyperrectangle
formulation, instead of seeing 'dimx' as the input dimension.